- implemented BOT_SERVER_GROUPS var - implemented uidIsBotClient HelperFunction - modified addOrRemoveSubscription to cope with track subscriptions

+++ sinusbot-event-subscriber.js
@@ -10,10 +10,15 @@ registerPlugin({
                 type: "select",
                 options: ["ERROR", "WARNING", "INFO", "VERBOSE"],
                 default: "2"
+            },
+            {
+                name: "BOT_SERVER_GROUPS",
+                title: "Bot Server Group Ids",
+                type: "strings"
             }
         ]
     },
-    function (_, {DEBUG_LEVEL}, {version}) {
+    function (_, {DEBUG_LEVEL, BOT_SERVER_GROUPS}, {version}) {
         const engine = require('engine');
         const store = require('store');
         const backend = require('backend');
@@ -270,6 +275,23 @@ registerPlugin({
                 return user.length > 0 ? user[0].name() : undefined;
             },
             /**
+             * Returns if a uid belongs to a bot client it the bot client is online
+             * @param {string}uid
+             * @returns {boolean}
+             */
+            uidIsBotClient: (uid) => {
+                const user = backend.getClientByUID(uid);
+                if (user != null) {
+                    let serverGroups = user.getServerGroups();
+                    serverGroups = serverGroups.filter(value => BOT_SERVER_GROUPS.includes(value.id()));
+                    if (serverGroups.length > 0) {
+                        return true;
+                    }
+                } else {
+                    return false;
+                }
+            },
+            /**
              * Add or Removes a subscription from the storage and replies to the client
              * @param client
              * @param args
@@ -278,6 +300,7 @@ registerPlugin({
              */
             addOrRemoveSubscription: (client, args, reply, sub) => {
                 let uid, nickname;
+                let commandPossible = false;
                 if (args.targetUId !== undefined && args.targetUId !== "" && args.targetNickname !== undefined && args.targetNickname !== "") {
                     reply("You should provide a targetNickname or a targetUId.");
                     reply("You can use a Nickname when your target is online or the uid if your target is offline.");
@@ -285,12 +308,21 @@ registerPlugin({
                 if ((args.targetUId === undefined || args.targetUId === "") && args.targetNickname !== undefined && args.targetNickname !== "") {
                     uid = args.targetNickname !== Subscription.allUId() ? HelperFunctions.nicknameToUId(args.targetNickname) : Subscription.allUId();
                     nickname = args.targetNickname;
+                    if (args.event.toUpperCase() === EventType.TRACK) {
+                        commandPossible = HelperFunctions.uidIsBotClient(uid);
+                    }
+                    commandPossible = true;
                 }
                 if (args.targetUId !== undefined && args.targetUId !== "" && (args.targetNickname === undefined || args.targetNickname === "")) {
                     uid = args.targetUId;
                     nickname = HelperFunctions.uidToNickname(uid);
+                    commandPossible = true;
                 }
-                if (uid !== undefined) {
+                //If Event is Track check if target is bot
+                if (commandPossible && args.event.toUpperCase() === EventType.TRACK) {
+                    commandPossible = HelperFunctions.uidIsBotClient(uid);
+                }
+                if (uid !== undefined && commandPossible) {
                     try {
                         let subscription = new SubscriptionBuilder().setSubscriberUId(client.uid()).setEvent(args.event.toUpperCase()).setTargetUId(uid).build();
                         subscription = sub ? subscriptionStore.addSubscription(subscription) : subscriptionStore.deleteSubscription(subscription);
